refactor: module on streaming
